Export MarshalGrid.
For the moment multiple controllers will need to push to the grid state.
This code is also non-trivial since grid state is compressed at rest.
Enable dependencies to reuse this code by exporting it.
