+++ pkg/updater/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "//pb/test_status:go_default_library",
         "//util/gcs:go_default_library",
         "@com_github_fvbommel_sortorder//:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_bazel_rules_go//proto/wkt:timestamp_go_proto",
         "@org_golang_google_api//googleapi:go_default_library",
@@ -46,7 +45,6 @@ go_test(
         "//util/gcs:go_default_library",
         "//util/gcs/fake:go_default_library",
         "@com_github_fvbommel_sortorder//:go_default_library",
-        "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
