+++ util/gcs/BUILD.bazel
@@ -37,6 +37,8 @@ go_test(
     deps = [
         "//metadata:go_default_library",
         "//metadata/junit:go_default_library",
+        "//pb/state:go_default_library",
+        "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
