+++ util/gcs/gcs_test.go
@@ -18,7 +18,11 @@ package gcs
 
 import (
 	"net/url"
+	"reflect"
 	"testing"
+
+	statepb "github.com/GoogleCloudPlatform/testgrid/pb/state"
+	"github.com/golang/protobuf/proto"
 )
 
 func Test_SetURL(t *testing.T) {
@@ -140,4 +144,36 @@ func Test_calcCRC(t *testing.T) {
 		t.Errorf("different results: %s %d != %s %d", b1, c1, b1a, c1a)
 	}
 
+}
+
+func TestMarshalGrid(t *testing.T) {
+	g1 := statepb.Grid{
+		Columns: []*statepb.Column{
+			{Build: "alpha"},
+			{Build: "second"},
+		},
+	}
+	g2 := statepb.Grid{
+		Columns: []*statepb.Column{
+			{Build: "first"},
+			{Build: "second"},
+		},
+	}
+
+	b1, e1 := MarshalGrid(&g1)
+	b2, e2 := MarshalGrid(&g2)
+	uncompressed, e1a := proto.Marshal(&g1)
+
+	switch {
+	case e1 != nil, e2 != nil:
+		t.Errorf("unexpected error %v %v %v", e1, e2, e1a)
+	}
+
+	if reflect.DeepEqual(b1, b2) {
+		t.Errorf("unexpected equality %v == %v", b1, b2)
+	}
+
+	if reflect.DeepEqual(b1, uncompressed) {
+		t.Errorf("should be compressed but is not: %v", b1)
+	}
 }
